changed log macros, changed LogInfo for less processing power, changed Thread to use pthread instead of c++ thread id, added two loggers

+++ util/src/Num.cpp
@@ -3,7 +3,7 @@
 namespace sihd::util
 {
 
-size_t  Num::get_size(ssize_t number, uint16_t base)
+size_t  Num::get_size(int64_t number, uint16_t base)
 {
     size_t ret = 0;
     while (number != 0)
